rework of double precision mode, now optional

+++ petrify-compiler-model/src/main/java/com/github/exabrial/petrify/compiler/model/LeafClassEntry.java
@@ -1,4 +1,4 @@
 package com.github.exabrial.petrify.compiler.model;
 
-public record LeafClassEntry(int classId, float weight) {
+public record LeafClassEntry(int classId, double weight) {
 }
