+++ petrify-compiler-model/src/main/java/com/github/exabrial/petrify/compiler/model/LeafTargetEntry.java
@@ -1,4 +1,4 @@
 package com.github.exabrial.petrify.compiler.model;
 
-public record LeafTargetEntry(int targetId, float weight) {
+public record LeafTargetEntry(int targetId, double weight) {
 }
